+++ client/package.json
@@ -8,8 +8,10 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.8",
     "apexcharts": "^3.41.1",
     "axios": "^1.4.0",
+    "bootstrap": "^5.3.1",
     "core-js": "^3.8.3",
     "vue": "^3.2.13",
     "vue3-apexcharts": "^1.4.4"
@@ -26,7 +28,8 @@
   "eslintConfig": {
     "root": true,
     "env": {
-      "node": true
+      "node": true,
+      "vue/setup-compiler-macros": true
     },
     "extends": [
       "plugin:vue/vue3-essential",
