+++ client/src/App.vue
@@ -1,83 +1,35 @@
 <template>
-  <apexchart width="1000" type="area" :options="options" :series="series" :key="chartKey"></apexchart>
+  <div class="row d-flex justify-content-center">
+    <div class="col-md-8 m-5">
+      <nav>
+        <div class="nav nav-tabs" id="nav-tab" role="tablist">
+          <template v-for="(res, index) in resources" :key="res.id">
+            <button class="nav-link" :id="'tab-' + index" data-bs-toggle="tab" :data-bs-target="'#nav-' + index" type="button" role="tab">{{res.name}} {{ index }}</button>
+          </template>
+        </div>
+      </nav>
+      <div class="tab-content" id="nav-tabContent">
+        <template v-for="(res, index) in resources" :key="res.id">
+          <div :class="'tab-pane fade ' + index == 0 ? 'show active' : ''" :id="'nav-' + index" role="tabpanel"><AggregateLineChart :resourceIndex="resourceIndex" />{{ index }}</div>
+        </template>
+      </div>  
+    </div>
+  </div>
+  
 </template>
 
 <script setup>
   import { ref, onMounted } from 'vue';
   import axios from 'axios';
-  import de from 'apexcharts/dist/locales/de.json';
+  import AggregateLineChart from './components/AggregateLineChart.vue';
 
-  const chartKey = ref(0);
-  let options = {};
-  let series = [];
+  const resourceIndex = ref(0);
+  const resources = ref([]);
 
   onMounted(async () => {
-    const chartData = await axios.get(process.env.VUE_APP_SERVER_URI + '/publicData/allConsumption');
-    const resourceTypes = await axios.get(process.env.VUE_APP_SERVER_URI + '/publicData/resourceTypes');
-
-    options = {
-      chart: {
-        id: "Aggregate consumption chart",
-        locales: [de],
-        defaultLocale: 'de'
-      },
-      title: {
-        text: "Aggregierter Ressourcenverbrauch pro Tag"
-      },
-      xaxis: {
-        categories: chartData.data[0].days,
-        type: 'datetime'
-      },
-      yaxis: {
-        labels: {
-          formatter: (value) => {
-            return value.toFixed(2);
-          }
-        },
-        tickAmount: 6,
-        title: {
-          text: resourceTypes.data[0].unit + "/Tag"
-        }
-      },
-      dataLabels: {
-        enabled: false
-      },
-      stroke: {
-        curve: 'smooth'
-      },
-      grid: {
-        row: {
-          colors: ['#eeeeee', 'transparent'],
-          opacity: 0.2
-        }
-      },
-      fill: {
-        type: "gradient",
-        gradient: {
-          shadeIntensity: 1,
-          opacityFrom: 0.7,
-          opacityTo: 0.9,
-          stops: [0, 90, 100]
-        }
-      },
-      tooltip: {
-        y: {
-          formatter: (value) => {
-            return value.toFixed(2);
-          }
-        }
-      }
-    };
-
-    series = [{
-      name: 'Täglicher Verbrauch',
-      data: chartData.data[0].consumption
-    }];
-
-    // Force Vue to re-render the chart (as no refs have been changed, there is no trigger otherwise)
-    chartKey.value +=1;
+    const response = await axios.get(process.env.VUE_APP_SERVER_URI + '/publicData/resourceTypes');
+    resources.value = response.data;
   });
-
 </script>
 
 <style>
